+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # MBS Changelog
 
+## v1.10.23
+* Drop BC R130 support
+* Add wheel of fortune support for the new R131 item positioning properties
+* Add userscript support for the newest BC domain
+
 ## v1.10.22
 * Add BC R131Beta2 support
 * Drop support for BC R131Beta1. Please upgrade to BC R131Beta2 or downgrade to R130